requirements.txt
discord.py
dhooks==1.1.4
requests==2.25.1
psutil==5.8.0
browser_cookie3==0.12.0
cryptography==3.4.7
pycryptodome==3.10.1